+++ asyncpraw/models/reddit/removal_reasons.py
@@ -4,6 +4,7 @@ from warnings import warn
 
 from ...const import API_PATH
 from ...exceptions import ClientException
+from ..util import deprecate_lazy
 from .base import RedditBase
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -144,13 +145,14 @@ class RemovalReason(RedditBase):
 class SubredditRemovalReasons:
     """Provide a set of functions to a Subreddit's removal reasons."""
 
+    @deprecate_lazy
     async def get_reason(
-        self, reason_id: Union[str, int, slice], lazy: bool = False
+        self, reason_id: Union[str, int, slice], fetch: bool = True, **kwargs
     ) -> RemovalReason:
         """Return the Removal Reason with the ID/number/slice ``reason_id``.
 
-        :param reason_id: The ID or index of the removal reason
-        :param lazy: If True, object is loaded lazily (default: False).
+        :param reason_id: The ID or index of the removal reason.
+        :param fetch: Determines if Async PRAW will fetch the object (default: True).
 
         This method is to be used to fetch a specific removal reason, like so:
 
@@ -195,7 +197,7 @@ class SubredditRemovalReasons:
 
             reason_id = "141vv5c16py7d"
             subreddit = await reddit.subreddit("NAME")
-            reason = await subreddit.mod.removal_reasons.get_reason(reason_id, lazy=True)
+            reason = await subreddit.mod.removal_reasons.get_reason(reason_id, fetch=False)
             await reason.delete()
 
         """
@@ -204,7 +206,7 @@ class SubredditRemovalReasons:
             return reasons[reason_id]
         else:
             reason = RemovalReason(self._reddit, self.subreddit, reason_id)
-        if not lazy:
+        if fetch:
             await reason._fetch()
         return reason