+++ tests/unit/test_deprecations.py
@@ -17,15 +17,6 @@ pytestmark = pytest.mark.filterwarnings(
 
 @pytest.mark.filterwarnings("error", category=DeprecationWarning)
 class TestDeprecation(UnitTest):
-    def test_validate_on_submit(self):
-        with pytest.raises(DeprecationWarning):
-            self.reddit.validate_on_submit
-        self.reddit.validate_on_submit = True
-        assert self.reddit.validate_on_submit
-        self.reddit.validate_on_submit = False
-        with pytest.raises(DeprecationWarning):
-            self.reddit.validate_on_submit
-
     def test_api_exception(self):
         exc = APIException(["test", "testing", "test"])
         with pytest.raises(DeprecationWarning):
@@ -34,6 +25,29 @@ class TestDeprecation(UnitTest):
             exc.message
         with pytest.raises(DeprecationWarning):
             exc.field
+
+    async def test_gild_method(self):
+        with pytest.raises(DeprecationWarning) as excinfo:
+            submission = await self.reddit.submission("1234", fetch=False)
+            await submission.gild()
+            assert excinfo.value.args[0] == "`.gild` has been renamed to `.award`."
+
+    def test_gold_method(self):
+        with pytest.raises(DeprecationWarning) as excinfo:
+            self.reddit.subreddits.gold()
+            assert (
+                excinfo.value.args[0]
+                == "`subreddits.gold` has be renamed to `subreddits.premium`."
+            )
+
+    async def test_lazy_argument_rename(self):
+        with pytest.deprecated_call() as warning_info:
+            await self.reddit.submission("1234", lazy=True)
+        assert (
+            str(warning_info.list[0].message)
+            == "The parameter ``lazy`` has been renamed to ``fetch`` and support for"
+            " the ``lazy`` parameter will be removed in a future version of Async PRAW."
+        )
 
     def test_praw_exception_rename(self):
         with pytest.raises(AsyncPRAWException):
@@ -52,42 +66,6 @@ class TestDeprecation(UnitTest):
         with pytest.raises(DeprecationWarning):
             from asyncpraw.exceptions import PRAWException  # noqa: F401
 
-    async def test_subreddit_rules_call(self):
-        with pytest.raises(DeprecationWarning) as excinfo:
-            subreddit = Subreddit(self.reddit, display_name="test")
-            await subreddit.rules()
-        assert (
-            excinfo.value.args[0]
-            == "Calling SubredditRules to get a list of rules is deprecated. Remove the parentheses to use the iterator. View the Async PRAW documentation on how to change the code in order to use the iterator (https://asyncpraw.readthedocs.io/en/latest/code_overview/other/subredditrules.html#asyncpraw.models.reddit.rules.SubredditRules.__call__)."
-        )
-
-    def test_web_socket_exception_attribute(self):
-        exc = WebSocketException("Test", Exception("Test"))
-        with pytest.raises(DeprecationWarning) as excinfo:
-            _ = exc.original_exception
-        assert (
-            excinfo.value.args[0]
-            == "Accessing the attribute original_exception is deprecated. Please rewrite your code in such a way that this attribute does not need to be used. It will be removed in Async PRAW 8.0."
-        )
-
-    def test_gold_method(self):
-        with pytest.raises(DeprecationWarning) as excinfo:
-            self.reddit.subreddits.gold()
-            assert (
-                excinfo.value.args[0]
-                == "`subreddits.gold` has be renamed to `subreddits.premium`."
-            )
-
-    async def test_gild_method(self):
-        with pytest.raises(DeprecationWarning) as excinfo:
-            submission = await self.reddit.submission("1234", lazy=True)
-            await submission.gild()
-            assert excinfo.value.args[0] == "`.gild` has been renamed to `.award`."
-
-    async def test_reddit_user_me_read_only(self):
-        with pytest.raises(DeprecationWarning):
-            await self.reddit.user.me()
-
     async def test_reddit_token_manager(self):
         with pytest.raises(DeprecationWarning):
             async with Reddit(
@@ -99,13 +77,31 @@ class TestDeprecation(UnitTest):
             ) as reddit:
                 reddit._core._requestor._http = None
 
+    async def test_reddit_user_me_read_only(self):
+        with pytest.raises(DeprecationWarning):
+            await self.reddit.user.me()
+
+    async def test_subreddit_rules_call(self):
+        with pytest.raises(DeprecationWarning) as excinfo:
+            subreddit = Subreddit(self.reddit, display_name="test")
+            await subreddit.rules()
+        assert (
+            excinfo.value.args[0]
+            == "Calling SubredditRules to get a list of rules is deprecated. Remove the"
+            " parentheses to use the iterator. View the Async PRAW documentation on how"
+            " to change the code in order to use the iterator"
+            " (https://asyncpraw.readthedocs.io/en/latest/code_overview/other/subredditrules.html#asyncpraw.models.reddit.rules.SubredditRules.__call__)."
+        )
+
     async def test_synchronous_context_manager(self):
         with pytest.raises(DeprecationWarning) as excinfo:
             with self.reddit:
                 pass
             assert (
                 excinfo.value.args[0]
-                == "Using this class as a synchronous context manager is deprecated and will be removed in the next release. Use this class as an asynchronous context manager instead."
+                == "Using this class as a synchronous context manager is deprecated and"
+                " will be removed in the next release. Use this class as an"
+                " asynchronous context manager instead."
             )
 
     def test_user_subreddit_as_dict(self):
@@ -115,10 +111,34 @@ class TestDeprecation(UnitTest):
             assert display_name == "test"
             assert (
                 warning_info.list[0].message.args[0]
-                == "`Redditor.subreddit` is no longer a dict and is now an `UserSubreddit` object. Accessing attributes using string indices is deprecated."
+                == "`Redditor.subreddit` is no longer a dict and is now an"
+                " `UserSubreddit` object. Accessing attributes using string indices is"
+                " deprecated."
             )
             assert user_subreddit.keys() == user_subreddit.__dict__.keys()
             assert (
                 warning_info.list[1].message.args[0]
-                == "`Redditor.subreddit` is no longer a dict and is now an `UserSubreddit` object. Using `keys` is deprecated and will be removed in Async PRAW 8."
+                == "`Redditor.subreddit` is no longer a dict and is now an"
+                " `UserSubreddit` object. Using `keys` is deprecated and will be"
+                " removed in Async PRAW 8."
             )
+
+    def test_validate_on_submit(self):
+        with pytest.raises(DeprecationWarning):
+            self.reddit.validate_on_submit
+        self.reddit.validate_on_submit = True
+        assert self.reddit.validate_on_submit
+        self.reddit.validate_on_submit = False
+        with pytest.raises(DeprecationWarning):
+            self.reddit.validate_on_submit
+
+    def test_web_socket_exception_attribute(self):
+        exc = WebSocketException("Test", Exception("Test"))
+        with pytest.raises(DeprecationWarning) as excinfo:
+            _ = exc.original_exception
+        assert (
+            excinfo.value.args[0]
+            == "Accessing the attribute original_exception is deprecated. Please"
+            " rewrite your code in such a way that this attribute does not need to"
+            " be used. It will be removed in Async PRAW 8.0."
+        )