+++ CHANGES.rst
@@ -21,6 +21,11 @@ Unreleased
 - An import error when using Async PRAW in environments where ``libsqlite3-dev`` is
   needed to utilize ``aiosqlite`` package which depends on the ``sqlite3`` builtin.
 
+**Deprecated**
+
+- The keyword argument ``lazy`` has been replace by ``fetch`` to consolidate the keyword
+  argument used to explicitly perform a fetch when initializing an object.
+
 7.4.0 (2021/07/30)
 ------------------